Re-enable role guessing, move to pedigrees package
