+++ dae/dae/pedigrees/tests/fixtures/sample_nuc_family.ped
